hoverCallback: fourth argument with row details;
Hover didn't have any clue about data point
position in the chart. This made it impossible
to move the "bubble label" near to the data
point which is actually hovered on.

This commit solved it by introducing an additional
hoverCallback function argument. That argument
differs from `content.data[index]' in the way that
it includes also the internal information about
canvas coordinates of the specific data point.

Now it is possible to make a hoverCallback like this:

function (index, options, content, row) {
  $('#'+options.element+' .morris-hover').css ({
    'position': 'absolute',
    'top':  row._y [0] - 35,
    'left': row._x - 35
  });
  return row.y;
}